// src/app/models/conceptDraw.model.ts
export class ConceptDrawModel {
  id?: number;
  code?: string;
  name?: string;
  preparedByClientId?: number;
  preparedByUserId?: number;
  clientId?: number;
  createdUserId?: number;
  updatedUserId?: number;
  typeProjectId?: number;
  typeApplicationId?: number;
  version?: number;
  approvedByClientId?: number;
  approvedByUserId?: number;
  xmlBase?: string;
  xmlAutocad?: string;
  csvQuote?: string;
  statusId?: number;
  userId?: number;
  clientName?: string;
  typeProject?: string;
  typeApplication?: string;
} 